feat(sensor): improved debouncing
was bouncing when reed switch released... a capacitor would be the best solution.

tried debouncing on both high and low, but was still seeing intermitant early reed switch state change, hence the speed based ignoring of state change.

the logic will only skip 4 intervals before using the new interval... that way if there is a sudden gust of wind, after 4 roations we'll acknowledge that the speed really has dramatically incrased